+++ BALSAMIC/utils/constants.py
@@ -538,6 +538,25 @@ REFERENCE_FILES = {
             "output_file": "delly_exclusion.tsv",
             "output_path": "genome"
         },
+        "ascat_gccorrection": {
+            "url":
+                "https://github.com/Clinical-Genomics/reference-files/blob/12a6c760fd542c02de2cda286b6245e46f4b6a97/cancer/references/GRCh37_SnpGcCorrections.tsv.gz",
+            "file_type": "text",
+            "gzip": True,
+            "genome_version": "hg19",
+            "output_file": "GRCh37_SnpGcCorrections.tsv",
+            "output_path": "genome"
+        },
+        "ascat_chryloci": {
+            "url":
+                "https://github.com/Clinical-Genomics/reference-files/raw/12a6c760fd542c02de2cda286b6245e46f4b6a97/cancer/references/GRCh37_d5_Y.loci",
+            "file_type": "text",
+            "gzip": False,
+            "genome_version": "hg19",
+            "output_file": "GRCh37_Y.loci",
+            "output_path": "genome"
+        },
+
     }
 }