+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Added:
 * ``ascatngs`` as a submodule
 * GitHub action to build and test ``ascatngs`` container
 * Reference section to docs/FAQ.rst
+* ascat download references
 * Delly tumor only rule
 
 Changed: